refactor: rename store -> stores
