+++ packages/vlossom/src/components/vs-toast/__tests__/vs-toast-view.test.ts
@@ -1,7 +1,7 @@
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 import { nextTick } from 'vue';
 import { mount } from '@vue/test-utils';
-import { store } from '@/store';
+import { store } from '@/stores';
 import VsToastView from '../VsToastView.vue';
 
 import type { ToastInfo } from '@/plugins';
