+++ packages/vlossom/src/composables/__tests__/color-scheme-composable.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { ref } from 'vue';
-import { store } from '@/store';
+import { store } from '@/stores';
 import { VsComponent } from '@/declaration';
 import { useColorScheme } from '../color-scheme-composable';
